Move `SignatureAlgorithm`
